Use DecimalField for threshold #16

+++ app/experimenter/experiments/models.py
@@ -30,7 +30,7 @@ class ExperimentVariant(models.Model):
         max_length=255, unique=True, blank=False, null=False)
     is_control = models.BooleanField(default=False)
     description = models.TextField(default='')
-    threshold = models.PositiveIntegerField(default=0)
+    threshold = models.DecimalField(max_digits=6, decimal_places=4, default=0.0)
     value = JSONField(default=False)
 
     def __str__(self):  # pragma: no cover
